Made data handling changes and refactored
