+++ src/Components/Layouts/Layout.js
@@ -2,6 +2,7 @@ import Header from "../Header/Header";
 import Footer from "../Footer/Footer";
 import NavBar from "../NavBar/NavBar";
 import { Outlet } from "react-router-dom";
+import "../NavBar/NavBar.css";
 
 const Layout = () => {
   return (
