+++ src/Components/NavBar/NavBar.css
@@ -11,7 +11,8 @@
     border-bottom: 1px solid LightGrey;
     border-top: 1px solid LightGrey;
     color: DarkGrey;
-
+    
+    
   }
 
 .link {
@@ -23,4 +24,10 @@
 
 .link:hover {
     opacity: 0.5;
+}
+
+.Navbar{
+    position:relative;
+    height: 100%;
+    overflow: auto;
 }
