--- src/Pages/Announcements.js
@@ -1,15 +1,0 @@
-
-import List from "./AnnouncementsList";
-import "./Announcements.css";
-
-
-const Announcements = () => {
-    return (
-      <div>
-        <h1>Announcements</h1>
-       <List/>
-      </div>
-    )
-  };
-  
-  export default Announcements;
