+++ src/Pages/Announcements/Styles/Announcements.css
@@ -1,8 +1,9 @@
 .Titel h2 {
     font-style: italic;
-    color: blue;
+    color: darkblue;
     padding-top: 0px;
     padding-bottom: 0px;
+    padding-left: 20px;
     margin-top: 0px;
     margin-bottom: 0px;
     margin-left: 0px;
@@ -10,6 +11,7 @@
 
 .Time p{
     font-style: italic;
+    font-size: small;
     color: gray;
     margin-bottom: 0px;
     margin-top: 10px;
@@ -21,6 +23,7 @@ p {
     margin-left: 20px;
     margin-top: 0px;
     margin-bottom: 0px;
+    margin-right: 20px;
 }
 
 h1 {
@@ -29,13 +32,16 @@ h1 {
 }
 
 .Border {
-    border: 1px dotted grey;
+    border: 1px solid grey;
     margin: 20px 20px 20px 20px;
     border-radius: 20px;
     padding-bottom: 20px;
+    opacity: 0.7;
+    max-width: 50%;
 }
 
 .Border:hover {
-    opacity: 0.5;
+    opacity: 1;
 }
+
